chore: Move all NS_ENUMs to NCTypes.h
