+++ NextcloudTalk/Calls/CallKitManager.m
@@ -6,7 +6,6 @@
 #import "CallKitManager.h"
 #import <CallKit/CXError.h>
 
-#import "CallConstants.h"
 #import "NCAudioController.h"
 #import "NCAPIController.h"
 #import "NCAppBranding.h"
