+++ NextcloudTalk/Calls/NCCallController.m
@@ -20,7 +20,6 @@
 
 #import "ARDCaptureController.h"
 
-#import "CallConstants.h"
 #import "CallKitManager.h"
 #import "NCAPIController.h"
 #import "NCAudioController.h"
