+++ NextcloudTalk/NextcloudTalk-Bridging-Header.h
@@ -74,7 +74,6 @@
 #import "MessageBodyTextView.h"
 #import "NCAudioController.h"
 #import "NCCallController.h"
-#import "CallConstants.h"
 #import "AvatarBackgroundImageView.h"
 #import "NCUser.h"
 #import "ContactsTableViewCell.h"
